[REFACTOR] test 코드 분리&이동
